Moves component to more appropriate place
